refactor(windowrule): consolidate exclusion-rule helpers into one header
Splitting "list → Exclude rules" (ExclusionListBridge.h) from "rules
→ Exclude slice" (ExcludeRuleFilter.h) left two LGPL headers doing
overlapping work: both build / probe Exclude-action WindowRules,
both pull in MatchExpression + RuleAction, both live in the same
namespace neighbourhood. Future readers had to discover that the
right helper depended on whether they were starting from a string
list or from a WindowRuleSet.

Collapse into one header, ExclusionRules.h, in the
PhosphorWindowRule::ExclusionRules namespace. The retained surface:

- makeExclusionRule(field, pattern) — builds one terminal Exclude
  rule with a deterministic UUIDv5 id (same namespace constant the
  v4 migration uses so a rule built here and a rule written by the
  migration collide cleanly).
- toRuleSet(applications, windowClasses) — the legacy
  effect-flavour `Contains`-leaf builder. Survives only for the
  animation-page's separate exclusion lists, which haven't migrated
  to unified rules yet.
- excludeRulesFrom(set) — derives the Exclude-shaped slice of a
  unified rule set (used by SnapEngine via the daemon and by the
  KWin effect's drag-gate). Replaces ExcludeRuleFilter::same.
- applicationExcludePatternsFrom(set) — extracts AppId pattern
  strings from Application-subject Exclude rules (for the few
  callers still on a flat pattern-list path).
- ruleIsExclude(rule) — shared predicate.

toDaemonRuleSet (the AppId AppIdMatches list-flavoured builder)
retires entirely: SnapEngine now reads the Exclude slice through
the unified rule store, the v4 migration produces the same UUIDv5
ids inline, and there were no other callers.

Deleted: ExclusionListBridge.h, ExcludeRuleFilter.h,
test_exclusionlistbridge.cpp. Renamed and re-namespaced:
test_exclusionrules.cpp / TestExclusionRules. Snap-engine, daemon,
KWin effect, and tests rewired to the consolidated header.
