+++ libs/phosphor-windowrule/CMakeLists.txt
@@ -72,7 +72,7 @@ set(phosphorwindowrule_public_HDRS
     # as sources (IDE visibility, install-intent parity with the rest of
     # the public headers above).
     include/PhosphorWindowRule/ContextRuleBridge.h
-    include/PhosphorWindowRule/ExclusionListBridge.h
+    include/PhosphorWindowRule/ExclusionRules.h
     include/PhosphorWindowRule/IdentityKey.h
 )
